linkedin_jobscraper/requirements.txt
selenium
webdriver-manager
beautifulsoup4
pandas
python-dotenv


